+++ source_locations.go
@@ -17,6 +17,26 @@ const (
 	OSMTypeWay      OSMType = "way"
 )
 
+// Tags
+
+type JSONBMapSlice []map[string]string
+
+func (j *JSONBMapSlice) Scan(value interface{}) error {
+	if value == nil {
+		*j = nil
+		return nil
+	}
+	b, ok := value.([]byte)
+	if !ok {
+		return fmt.Errorf("unsupported type: %T", value)
+	}
+	return json.Unmarshal(b, j)
+}
+
+func (j JSONBMapSlice) Value() (driver.Value, error) {
+	return json.Marshal(j)
+}
+
 // Reports
 
 type ReportType string
